requirements/requirements.txt
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.19.0
protobuf==5.27.3
pandas==2.2.2
pathlib2==2.3.7.post1
PyYAML==6.0.2
scipy==1.13.1
tabulate==0.9.0
torcheval==0.0.7
torchinfo==1.8.0
cryptography==43.0.1
matplotlib==3.9.2
scikit-learn==1.5.2
pywavelets==1.8.0
